test(select0: migrate highlight
